requirements.txt
argcomplete==2.0.0
asgiref==3.5.2
boto3==1.24.80
botocore==1.27.80
certifi==2022.9.24
cffi==1.15.1
cfn-flip==1.3.0
charset-normalizer==2.1.1
click==8.1.3
crispy-bootstrap5==0.6
cryptography==38.0.1
defusedxml==0.7.1
Django==4.0.7
django-allauth==0.50.0
django-ckeditor==6.4.2
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-dotenv==1.4.2
django-filter==22.1
django-js-asset==2.0.0
django-s3-storage==0.13.10
django-storages==1.13.1
djangorestframework==3.14.0
durationpy==0.5
gunicorn==20.1.0
hjson==3.1.0
idna==3.4
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.1
oauthlib==3.2.1
Pillow==9.2.0
placebo==0.9.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.5.0
python-dateutil==2.8.2
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==6.0
qrcode==7.3.1
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
toml==0.10.2
tqdm==4.64.1
troposphere==4.1.0
types-cryptography==3.3.23
urllib3==1.26.12
Werkzeug==2.2.2
wsgi-request-logger==0.4.6
zappa==0.55.0
